Rename components and views
